Refactor to merge UserService and UserServiceImpl
