+++ src/main/java/com/codessquad/qna/user/UserService.java
@@ -1,15 +1,57 @@
 package com.codessquad.qna.user;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+
+import java.nio.charset.StandardCharsets;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
-public interface UserService {
-    List<UserDTO> getUsers();
+@Service
+public class UserService {
+    private final UserRepository userRepository;
 
-    UserDTO getUser(Long id);
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
-    UserDTO getUserWithVerifyPassword(String userId, String password);
+    public List<UserDTO> getUsers() {
+        return StreamSupport.stream(userRepository.findAll().spliterator(), true)
+                .map(User::toDTO)
+                .collect(Collectors.toList());
+    }
 
-    void createUser(UserDTO userDTO);
+    public UserDTO getUser(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자 입니다. id : " + id))
+                .toDTO();
+    }
 
-    UserDTO updateUser(Long id, UserDTO newUser);
+    public UserDTO getUserWithVerifyPassword(String userId, String password) {
+        return userRepository.findByUserId(userId)
+                .orElseThrow(() -> HttpClientErrorException.create(
+                        User.LOGIN_FAIL_MESSAGE,
+                        HttpStatus.UNAUTHORIZED,
+                        "",
+                        null,
+                        null,
+                        StandardCharsets.UTF_8
+                ))
+                .toDTO();
+    }
+
+    public void createUser(UserDTO userDTO) {
+        userRepository.save(userDTO.toEntity());
+    }
+
+    public UserDTO updateUser(Long id, UserDTO newUser) {
+        User existedUser = getUser(id).toEntity();
+
+        existedUser.checkPassword(newUser.getPassword());
+        existedUser.update(newUser);
+
+        return userRepository.save(existedUser).toDTO();
+    }
 }
